requirements.txt
alembic==1.11.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==3.2.0
billiard==4.2.1
black==23.9.1
celery==5.3.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.9
cryptography==44.0.0
Deprecated==1.2.15
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
flake8==6.1.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jwt==1.3.1
kombu==5.4.2
limits==3.14.1
loguru==0.7.0
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
mysqlclient==2.2.6
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==4.21.12
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyflakes==3.1.0
PyMySQL==1.1.0
pytest==7.4.2
pytest-asyncio==0.21.0
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.0
requests==2.31.0
rsa==4.9
setuptools==75.6.0
six==1.17.0
slowapi==0.1.8
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.22.0
vine==5.1.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.17.0
